% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMMA_comp.R
\name{export_remma_scores}
\alias{export_remma_scores}
\title{Export REMMA analysis scores to CSV}
\usage{
export_remma_scores(em_rob, out_path)
}
\arguments{
\item{em_rob}{object output from run_remma_analysis()$em_rob}

\item{out_path}{path to write the combined scores tibble}
}
\description{
Export REMMA analysis scores to CSV
}
