% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_heusser_pollen.R
\name{generate_heusser_pollen}
\alias{generate_heusser_pollen}
\title{Generate Heusser (1990) pollen count diagram and save output}
\usage{
generate_heusser_pollen(
  raw_path = here::here("analysis/data/raw_data"),
  output_dir = here::here("analysis/supplement")
)
}
\arguments{
\item{raw_path}{Path to folder containing the .tab and .csv files}

\item{output_dir}{Directory to save metadata, CSV, and plot (default: analysis/supplement)}
}
\description{
Generate Heusser (1990) pollen count diagram and save output
}
