% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_palynomorph_rioja.R
\name{generate_palynomorph_rioja}
\alias{generate_palynomorph_rioja}
\title{Generate palynomorph Rioja diagram and save output}
\usage{
generate_palynomorph_rioja(
  raw_path = here::here("analysis/data/raw_data"),
  output_dir = here::here("analysis/supplement")
)
}
\arguments{
\item{raw_path}{Path to folder containing \code{TTGpalynomorphsRioja.csv}}

\item{output_dir}{Directory to save plot (default: analysis/supplement)}
}
\description{
Generate palynomorph Rioja diagram and save output
}
