% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GechronR.R
\name{get_ensemble}
\alias{get_ensemble}
\title{Extract ensemble data for a given model}
\usage{
get_ensemble(lpd, model.num)
}
\arguments{
\item{lpd}{A LiPD object with run models added}

\item{model.num}{The model number (1 for Bacon, 2 for Bchron)}
}
\value{
A tibble with columns depth and ageEnsemble
}
\description{
Extract ensemble data for a given model
}
