% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_and_prep_data}
\alias{load_and_prep_data}
\title{Load and preprocess all raw data}
\usage{
load_and_prep_data(raw_path = system.file("testdata", package = "TT3"))
}
\arguments{
\item{raw_path}{character path to the raw_data folder}
}
\value{
A list with elements diatom, palyno, phyto, zone_data
}
\description{
Load and preprocess all raw data
}
