% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GechronR.R
\name{load_lpd}
\alias{load_lpd}
\title{Load a LiPD dataset from a folder}
\usage{
load_lpd(folder)
}
\arguments{
\item{folder}{Path to the LiPD .lpd folder}
}
\value{
A LiPD object
}
\description{
Load a LiPD dataset from a folder
}
