% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMMA_comp.R
\name{load_remma_data}
\alias{load_remma_data}
\title{Preprocess REMMA sediment and grain size data}
\usage{
load_remma_data(path_x0, path_tte)
}
\arguments{
\item{path_x0}{path to CSV file with H01 data}

\item{path_tte}{path to TTGmicro CSV (grain size matrix)}
}
\value{
A list with elements H01_data (tibble) and gs_matrix (matrix)
}
\description{
Preprocess REMMA sediment and grain size data
}
