% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GechronR.R
\name{plot_age_ensemble}
\alias{plot_age_ensemble}
\title{Plot ensemble age-depth ribbons for multiple models}
\usage{
plot_age_ensemble(ensembles, colors = c(Bacon = "red", Bchron = "blue"))
}
\arguments{
\item{ensembles}{A named list of tibbles from get_ensemble}

\item{colors}{A named vector of colors for each model}
}
\value{
A ggplot object
}
\description{
Plot ensemble age-depth ribbons for multiple models
}
