% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anomoeoneis.R
\name{plot_anomoeoneis}
\alias{plot_anomoeoneis}
\title{Plot Anomoeoneis sculpa abundance}
\usage{
plot_anomoeoneis(diatom_df, zone_data)
}
\arguments{
\item{diatom_df}{tibble from load_and_prep_data()$diatom}

\item{zone_data}{tibble from load_and_prep_data()$zone_data}
}
\value{
A ggplot2 object
}
\description{
Plot Anomoeoneis sculpa abundance
}
