% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diatomPlot.R
\name{plot_diatom_groups}
\alias{plot_diatom_groups}
\title{Plot diatom group abundances}
\usage{
plot_diatom_groups(diatom_data, zone_data)
}
\arguments{
\item{diatom_data}{tibble resultado de tu procesamiento (taxon.sel.lon)}

\item{zone_data}{tibble con las zonas a sombrear (ymin, ymax, xmin, xmax)}
}
\value{
ggplot object
}
\description{
Plot diatom group abundances
}
