% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diatom_pb.R
\name{plot_diatom_pb}
\alias{plot_diatom_pb}
\title{Plot Diatom P/B Index}
\usage{
plot_diatom_pb(diatom_df, zone_data)
}
\arguments{
\item{diatom_df}{tibble from load_and_prep_data()$diatom}

\item{zone_data}{tibble from load_and_prep_data()$zone_data}
}
\value{
A ggplot2 object
}
\description{
Plot Diatom P/B Index
}
