% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4.R
\name{plot_figure4}
\alias{plot_figure4}
\title{Combine all stratigraphy plots into Figure 4}
\usage{
plot_figure4(data_list)
}
\arguments{
\item{data_list}{output of load_and_prep_data()}
}
\value{
A patchwork object
}
\description{
Combine all stratigraphy plots into Figure 4
}
