% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palynomorph.R
\name{plot_palynomorph}
\alias{plot_palynomorph}
\title{Plot palynomorph groups}
\usage{
plot_palynomorph(palyno_df, zone_data)
}
\arguments{
\item{palyno_df}{tibble from load_and_prep_data()$palyno}

\item{zone_data}{tibble from load_and_prep_data()$zone_data}
}
\value{
A ggplot2 object
}
\description{
Plot palynomorph groups
}
