% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phytolith.R
\name{plot_phytolith}
\alias{plot_phytolith}
\title{Plot phytolith groups}
\usage{
plot_phytolith(phyto_df, zone_data)
}
\arguments{
\item{phyto_df}{tibble from load_and_prep_data()$phyto}

\item{zone_data}{tibble from load_and_prep_data()$zone_data}
}
\value{
A ggplot2 object
}
\description{
Plot phytolith groups
}
