% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GechronR.R
\name{run_age_models}
\alias{run_age_models}
\title{Run Bacon and Bchron age models}
\usage{
run_age_models(
  lpd,
  bacon_dir,
  bacon_args = list(bacon.thick = 3, lab.id.var = "labid", model.num = 1, age.var =
    "age", age.uncertainty.var = "ageuncertainty", depth.var = "depth",
    reservoir.age.14c.var = "dR", reservoir.age.14c.uncertainty.var = "dSTD",
    accept.suggestions = TRUE, cc = 3, ssize = 2000, d.min = 38, d.max = 266,
    bacon.acc.mean = c(70, 100), acc.shape = c(1.3, 1.6), mem.mean = 0.4, hiatus.depths =
    165, hiatus.max = 500, ask = FALSE),
  bchron_args = list(iter = 40000, model.num = 2, predictPositions = seq(38, 266, by =
    1), lab.id.var = "labid", age.var = "age", age.uncertainty.var = "ageuncertainty",
    depth.var = "depth", reservoir.age.14c.var = "dR", reservoir.age.14c.uncertainty.var
    = "dSTD", cal.curves = "shcal20")
)
}
\arguments{
\item{lpd}{A LiPD object}

\item{bacon_dir}{Directory to store Bacon output}

\item{bacon_args}{List of additional arguments for runBacon}

\item{bchron_args}{List of additional arguments for runBchron}
}
\value{
A LiPD object with added age models
}
\description{
Run Bacon and Bchron age models
}
