% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMMA_comp.R
\name{run_remma_analysis}
\alias{run_remma_analysis}
\title{Perform environmental mode analysis using EMMAgeo}
\usage{
run_remma_analysis(gs_matrix, max_corr = 0.95, n = 20, plot = TRUE)
}
\arguments{
\item{gs_matrix}{grain size matrix (rows=records, cols=phi classes)}

\item{max_corr}{numeric maximum correlation threshold}

\item{n}{number of potential EMs}

\item{plot}{logical, whether to plot intermediate results}
}
\value{
A list with EM potential model, robust loadings, and scores
}
\description{
Perform environmental mode analysis using EMMAgeo
}
