% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4.R
\name{save_figure4}
\alias{save_figure4}
\title{Save Figure 4 to file}
\usage{
save_figure4(filename, data_list)
}
\arguments{
\item{filename}{path where to save the image}

\item{data_list}{output of load_and_prep_data()}
}
\description{
Save Figure 4 to file
}
